refactor: move frontmatter.py into gax/gsheet/ package
Sheets-specific module, now lives with its consumers.
